backend/requirements.txt
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.2
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyMySQL==1.0.3
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
